feat: Add step 2 from launchpad page
This commit adds a new page for the admin order creation flow. This page
is responsible for initializing the order creation process with data
passed from the launchpad, specifically customer information. It
retrieves parameters from the URL, initializes the order creation state,
and then redirects the user to the next step in the order creation
process.

https://github.com/over55/monorepo/issues/55