feat: Add filters for initiator, open/closed status
Adds filtering capabilities to the order incident listing.

The following filters are added: - `initiator`: Filter by the initiator
of the incident. - `status`: Now accepts "open" and "closed" to filter
for open and closed incidents, respectively. Also allows for a numeric
status code for backwards compatibility.

Also fixes some bugs in the filtering logic.

https://github.com/over55/monorepo/issues/62