feat: Support public endpoint for S3 presigned URLs
This commit introduces a separate public endpoint configuration for
generating S3 presigned URLs. This is crucial for scenarios where
internal and external access points to the S3 service differ, such as in
Docker environments using minio. The existing S3 client is used for
internal operations, while a new client, configured with the public
endpoint, is used for generating presigned URLs that are accessible from
outside the service (e.g. browsers).

https://github.com/over55/monorepo/issues/40