feat: Add reason filter to away log datastore
This commit adds the `Reason` field to the `AssociateAwayLogListFilter`,
`AssociateAwayLogPaginationListFilter` and adds the `Reason` field to
the `CountByFilter` for the datastore. This allows filtering of away
logs by reason.

This also includes the addition of the `StatusActive` and
`StatusArchived` constants.

The logic in the `createAwayLogOnExpiredDuesDate` function was updated
to properly create and maintain the associate away logs.

https://github.com/over55/monorepo/issues/49