feat: Add filter by associate IDs for away logs
This commit adds the `InAssociateIDs` filter to the `AssociateAwayLog`
datastore and controller. This allows filtering of away logs by a list
of associate IDs. Also, it adds the `IsAway` flag to the
`AssignableAssociate` type to indicate if an associate is currently
away. This is used to determine if an associate can be assigned to a
task.

https://github.com/over55/monorepo/issues/80